[refactor #35] ObjectMapperBasedVoMapper 클래스 Common 모듈 이동 & 패키지 구조 변경
