+++ auth/src/main/java/com/ticketPing/auth/AuthApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"com.ticketPing.auth", "audit", "exception", "aop"})
+@ComponentScan(basePackages = {"com.ticketPing.auth", "audit", "common/exception", "common/aop"})
 public class AuthApplication {
 
 	public static void main(String[] args) {
