+++ auth/src/main/java/com/ticketPing/auth/presentation/controller/AuthController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import response.CommonResponse;
+import common.response.CommonResponse;
 
 @RestController
 @RequiredArgsConstructor
