+++ auth/src/main/java/com/ticketPing/auth/presentation/status/AuthErrorCase.java
@@ -1,6 +1,6 @@
 package com.ticketPing.auth.presentation.status;
 
-import cases.ErrorCase;
+import common.cases.ErrorCase;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
