+++ common/src/main/java/common/aop/LoggingAspect.java
@@ -1,4 +1,4 @@
-package aop;
+package common.aop;
 
 import org.aspectj.lang.JoinPoint;
 import org.aspectj.lang.annotation.After;
