+++ common/src/main/java/common/cases/SuccessCase.java
@@ -1,9 +1,8 @@
-package cases;
+package common.cases;
 
 import org.springframework.http.HttpStatus;
 
-public interface ErrorCase {
+public interface SuccessCase {
     HttpStatus getHttpStatus();
     String getMessage();
 }
-
