+++ common/src/main/java/common/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package exception;
+package common.exception;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -11,7 +11,7 @@ import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
-import response.CommonResponse;
+import common.response.CommonResponse;
 
 @ControllerAdvice
 @RequiredArgsConstructor
