+++ services/client-channel/src/main/java/com/ticketPing/client_channel/ClientChannelApplication.java
@@ -6,7 +6,7 @@ import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.ComponentScan;
 
 @EnableFeignClients
-@ComponentScan(basePackages = {"com.ticketPing.client_channel", "exception", "aop"})
+@ComponentScan(basePackages = {"com.ticketPing.client_channel", "common"})
 @SpringBootApplication
 public class ClientChannelApplication {
