+++ services/client-channel/src/main/java/com/ticketPing/client_channel/application/client/QueueManageClientAdapter.java
@@ -1,9 +1,10 @@
 package com.ticketPing.client_channel.application.client;
 
 import com.ticketPing.client_channel.infrastructure.annotation.FeignClientAdapter;
+import com.ticketPing.client_channel.infrastructure.client.QueueManageClient;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
-import response.CommonResponse;
+import common.response.CommonResponse;
 
 @FeignClientAdapter
 @RequiredArgsConstructor
