+++ services/client-channel/src/main/java/com/ticketPing/client_channel/infrastructure/client/QueueManageClient.java
@@ -1,11 +1,11 @@
-package com.ticketPing.client_channel.application.client;
+package com.ticketPing.client_channel.infrastructure.client;
 
 import com.ticketPing.client_channel.application.dto.WaitingQueueInfoResponse;
 import java.util.UUID;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import response.CommonResponse;
+import common.response.CommonResponse;
 
 @FeignClient(name = "queue-manage")
 public interface QueueManageClient {
