+++ services/order/src/main/java/com/ticketPing/order/presentation/response/exception/OrderExceptionCase.java
@@ -1,7 +1,7 @@
 package com.ticketPing.order.presentation.response.exception;
 
 
-import cases.ErrorCase;
+import common.cases.ErrorCase;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
