+++ services/order/src/main/java/com/ticketPing/order/presentation/response/success/OrderSuccessCase.java
@@ -1,6 +1,6 @@
 package com.ticketPing.order.presentation.response.success;
 
-import cases.SuccessCase;
+import common.cases.SuccessCase;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
