+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/application/service/WaitingQueueApplicationServiceImpl.java
@@ -1,9 +1,13 @@
 package com.ticketPing.queue_manage.application.service;
 
+import static com.ticketPing.queue_manage.domain.model.WaitingQueueToken.create;
+
 import com.ticketPing.queue_manage.application.dto.EnterWaitingQueueResponse;
 import com.ticketPing.queue_manage.application.dto.WaitingQueueInfoResponse;
+import com.ticketPing.queue_manage.domain.command.waitingQueue.EnqueueCommand;
+import com.ticketPing.queue_manage.domain.command.waitingQueue.RetrieveTokenCommand;
 import com.ticketPing.queue_manage.domain.model.WaitingQueueToken;
-import com.ticketPing.queue_manage.domain.service.WaitingQueueDomainService;
+import com.ticketPing.queue_manage.domain.repository.WaitingQueueRepository;
 import com.ticketPing.queue_manage.presentaion.request.EnterWaitingQueueRequest;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
@@ -13,17 +17,18 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class WaitingQueueApplicationServiceImpl implements WaitingQueueApplicationService {
 
-    private final WaitingQueueDomainService waitingQueueDomainService;
+    private final WaitingQueueRepository waitingQueueRepository;
 
     @Override
     public EnterWaitingQueueResponse enterWaitingQueue(EnterWaitingQueueRequest request) {
-        WaitingQueueToken token = waitingQueueDomainService.enterWaitingQueue(request);
+        WaitingQueueToken token = create(request);
+        waitingQueueRepository.enqueue(EnqueueCommand.create(token));
         return EnterWaitingQueueResponse.from(token);
     }
 
     @Override
     public WaitingQueueInfoResponse getWaitingQueueInfo(UUID userId) {
-        WaitingQueueToken token = waitingQueueDomainService.retrieveToken(userId);
+        WaitingQueueToken token = waitingQueueRepository.retrieveToken(RetrieveTokenCommand.create(userId));
         return WaitingQueueInfoResponse.from(token);
     }
