+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/application/service/WorkingQueueApplicationServiceImpl.java
@@ -1,6 +1,14 @@
 package com.ticketPing.queue_manage.application.service;
 
-import com.ticketPing.queue_manage.domain.service.WorkingQueueDomainService;
+import com.ticketPing.queue_manage.domain.command.waitingQueue.DequeueCommand;
+import com.ticketPing.queue_manage.domain.command.waitingQueue.RetrieveTopTokensCommand;
+import com.ticketPing.queue_manage.domain.command.workingQueue.CacheTokenCommand;
+import com.ticketPing.queue_manage.domain.command.workingQueue.CountAvailableSlotsCommand;
+import com.ticketPing.queue_manage.domain.command.workingQueue.IncrementCounterCommand;
+import com.ticketPing.queue_manage.domain.model.AvailableSlots;
+import com.ticketPing.queue_manage.domain.model.WorkingQueueToken;
+import com.ticketPing.queue_manage.domain.repository.WaitingQueueRepository;
+import com.ticketPing.queue_manage.domain.repository.WorkingQueueRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -8,11 +16,26 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 public class WorkingQueueApplicationServiceImpl implements WorkingQueueApplicationService {
 
-    private final WorkingQueueDomainService workingQueueDomainService;
+    private final WaitingQueueRepository waitingQueueRepository;
+    private final WorkingQueueRepository workingQueueRepository;
 
     @Override
     public void processQueueTransfer() {
-        workingQueueDomainService.processQueueTransfer();
+        // 작업열 인원 여유 확인
+        AvailableSlots availableSlots = workingQueueRepository.countAvailableSlots(CountAvailableSlotsCommand.create());
+        if (availableSlots.isLimited()) {
+            return;
+        }
+
+        // 여유 있으면 대기열 -> 작업열 이동
+        long count = availableSlots.getCount();
+        waitingQueueRepository.retrieveTopTokens(RetrieveTopTokensCommand.create(count)).stream()
+                .toList()
+                .forEach(waitingToken -> {
+                    WorkingQueueToken workingQueueToken = waitingToken.toWorkingQueueToken();
+                    workingQueueRepository.enqueue(CacheTokenCommand.create(workingQueueToken), IncrementCounterCommand.create());
+                    waitingQueueRepository.dequeue(DequeueCommand.create(waitingToken));
+                });
     }
 
 }
