+++ services/queue-manage/src/main/java/com/ticketPing/queue_manage/domain/repository/WaitingQueueRepositoryImpl.java
@@ -1,6 +1,6 @@
 package com.ticketPing.queue_manage.domain.repository;
 
-import static com.ticketPing.queue_manage.domain.cases.QueueErrorCase.USER_NOT_FOUND;
+import static com.ticketPing.queue_manage.presentaion.cases.QueueErrorCase.USER_NOT_FOUND;
 import static com.ticketPing.queue_manage.domain.model.WaitingQueueToken.tokenWithPosition;
 
 import com.ticketPing.queue_manage.domain.command.waitingQueue.DequeueCommand;
@@ -8,7 +8,7 @@ import com.ticketPing.queue_manage.domain.command.waitingQueue.EnqueueCommand;
 import com.ticketPing.queue_manage.domain.command.waitingQueue.RetrieveTokenCommand;
 import com.ticketPing.queue_manage.domain.command.waitingQueue.RetrieveTopTokensCommand;
 import com.ticketPing.queue_manage.domain.model.WaitingQueueToken;
-import exception.ApplicationException;
+import common.exception.ApplicationException;
 import java.util.Collection;
 import java.util.List;
 import java.util.stream.Collectors;
