--- services/queue-manage/src/main/java/com/ticketPing/queue_manage/domain/service/WorkingQueueDomainService.java
@@ -1,5 +1,0 @@
-package com.ticketPing.queue_manage.domain.service;
-
-public interface WorkingQueueDomainService {
-    void processQueueTransfer();
-}
