+++ services/user/src/main/java/com/ticketPing/user/UserApplication.java
@@ -5,7 +5,7 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
 
 @SpringBootApplication
-@ComponentScan(basePackages = {"com.ticketPing.user", "audit", "exception", "aop"})
+@ComponentScan(basePackages = {"com.ticketPing.user", "audit", "common/exception", "common/aop"})
 public class UserApplication {
 
 	public static void main(String[] args) {
